+++ webapi/src/main/scala/org/knora/webapi/slice/admin/api/Codecs.scala
@@ -108,7 +108,7 @@ object Codecs {
     implicit val sparqlEncodedString: StringCodec[SparqlEncodedString] = stringCodec(SparqlEncodedString.from)
     implicit val status: StringCodec[Status]                           = booleanCodec(Status.from)
     implicit val copyrightHolder: StringCodec[CopyrightHolder]         = stringCodec(CopyrightHolder.from)
-    implicit val licenseText: StringCodec[LicenseText]                 = stringCodec(LicenseText.from)
+    implicit val licenseIdentifier: StringCodec[LicenseIdentifier]     = stringCodec(LicenseIdentifier.from)
     implicit val licenseUri: StringCodec[LicenseUri]                   = stringCodec(LicenseUri.from)
 
     // user